+++ src/comentarios/comentarios.controller.ts
@@ -1,3 +1,4 @@
+import { CurrentUser } from './../auth/decorators/currentUser.decorator';
 import {
   Controller,
   Get,
@@ -9,11 +10,13 @@ import {
   NotFoundException,
   HttpCode,
   ParseIntPipe,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { ComentariosService } from './comentarios.service';
 import { CreateComentariosDto } from './dto/create-comentarios.dto';
 import { UpdateComentariosDto } from './dto/update-comentarios.dto';
 import { Public } from 'src/auth/decorators/isPublic.decorator';
+import { UserPayload } from 'src/auth/types/UserPayload';
 
 @Controller('comentarios')
 export class ComentariosController {
@@ -21,7 +24,13 @@ export class ComentariosController {
 
   @Post()
   @HttpCode(201) // Recurso criado com sucesso.
-  async create(@Body() createComentariosDto: CreateComentariosDto) {
+  async create(
+    @Body() createComentariosDto: CreateComentariosDto,
+    @CurrentUser() currentUser: UserPayload,
+  ) {
+    if (createComentariosDto.usuarioId !== currentUser.sub) {
+      throw new UnauthorizedException('Unauthorized.');
+    }
     return this.comentariosService.create(createComentariosDto);
   }
 
@@ -48,6 +57,7 @@ export class ComentariosController {
   async update(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateComentariosDto: UpdateComentariosDto,
+    @CurrentUser() currentUser: UserPayload,
   ) {
     const comentario = await this.comentariosService.findOne(id);
     if (!comentario) {
@@ -56,18 +66,27 @@ export class ComentariosController {
         `Comentário com ID ${id} não encontrado para atualização.`,
       );
     }
+    if (comentario.usuarioId !== currentUser.sub) {
+      throw new UnauthorizedException('Unauthorized.');
+    }
     return this.comentariosService.update(id, updateComentariosDto);
   }
 
   @Delete(':id')
   @HttpCode(204) // Recurso excluído com sucesso.
-  async remove(@Param('id', ParseIntPipe) id: number) {
+  async remove(
+    @Param('id', ParseIntPipe) id: number,
+    @CurrentUser() currentUser: UserPayload,
+  ) {
     const comentario = await this.comentariosService.findOne(id);
     if (!comentario) {
       // Lança erro 404 se o comentário não for encontrado.
       throw new NotFoundException(
         `Comentário com ID ${id} não encontrado para exclusão.`,
       );
+    }
+    if (comentario.usuarioId !== currentUser.sub) {
+      throw new UnauthorizedException('Unauthorized.');
     }
     return this.comentariosService.remove(id);
   }
